change view directory
