+++ nbproject/private/private.properties
@@ -1,2 +1,8 @@
 compile.on.save=true
-user.properties.file=C:\\Users\\42labinfo\\AppData\\Roaming\\NetBeans\\15\\build.properties
+do.depend=false
+do.jar=true
+do.jlink=false
+javac.debug=true
+javadoc.preview=true
+jlink.strip=false
+user.properties.file=C:\\Users\\Lucas Dantas\\AppData\\Roaming\\NetBeans\\17\\build.properties
