+++ nbproject/project.properties
@@ -1,9 +1,10 @@
 annotation.processing.enabled=true
 annotation.processing.enabled.in.editor=false
-annotation.processing.processor.options=
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
+application.title=projetoIntegrador
+application.vendor=Lucas Dantas
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
 # This directory is removed when the project is cleaned:
@@ -32,10 +33,14 @@ dist.jar=${dist.dir}/projetoIntegrador.jar
 dist.javadoc.dir=${dist.dir}/javadoc
 dist.jlink.dir=${dist.dir}/jlink
 dist.jlink.output=${dist.jlink.dir}/projetoIntegrador
+endorsed.classpath=
 excludes=
+file.reference.mysql-connector-java-8.0.30.jar=mysql-connector-java-8.0.30.jar
 includes=**
 jar.compress=false
-javac.classpath=
+javac.classpath=\
+    ${libs.absolutelayout.classpath}:\
+    ${file.reference.mysql-connector-java-8.0.30.jar}
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
@@ -71,7 +76,7 @@ jlink.additionalmodules=
 jlink.additionalparam=
 jlink.launcher=true
 jlink.launcher.name=projetoIntegrador
-main.class=projetointegrador.ProjetoIntegrador
+main.class=views.JavaEstoque
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
